requirements.txt
aerich
aiofiles
aiogram
aiogram-media-group
aiogram3-form
aiohttp
aioschedule
aiosignal
aiosqlite
async-timeout
attrs
certifi
charset-normalizer
click
colorama
dictdiffer
frozenlist
idna
iso8601
magic-filter
multidict
pydantic
Pygments
pypika-tortoise
pytz
tomlkit
tortoise-orm
typer
typing_extensions
yarl
geopy
aiogram-calendar
apscheduler
aiomysql